Add Clearfix
Clearfix is an extremely helpful bit of code to clear floating divs in
any browser. In my opinion it is as essential as a Reset. I've
implemented here utilizing the Skeleton IE classes assigned in the
index.html file rather than other popular css hacks often used for IE
targeting on clearfix.