Improve anti-aliasing on WebKit browsers
`-webkit-font-smoothing` provides WebKit browsers with improved anti-aliasing capabilities. The reason to nest this with the ever-notorious universal identifier (`*`) is that not HTML elements are caught under the `body` blanket. This is most notable when comparing the way fonts render text inside of `<input[type="button"]>` form elements as opposed to purely textual elements like `<a>`, `<p>`. To deliver a beautiful, consistent experience, we need to apply this universally for all elements on WebKit.