Fixed inconsistent offsets
Offsets are given a width of (x columns + one margin). When applying an
offset to a column that is positioned to the right of another column,
the offset is too short, as the first column has no margin. This commit
changes the direction of margins, so that each column can properly pad
offsets.
Also I removed a redundant width: 100%.
margin-right: 0; may also be redundant, but it may be a browserhack so I
won't remove it.